matching python package tree
